requirements.txt
bernard
nasa-api-wrapper==0.1.5
pendulum==2.0.4
tqdm==4.32.1
